+++ src/DriverInterface.php
@@ -9,9 +9,11 @@ use Closure;
 interface DriverInterface
 {
     /**
+     * @param Closure $onResolve called on resolved and first argument is $callback return or Flow\Exception on Exception
+     *
      * @return Closure when called, this start async $callback
      */
-    public function async(Closure $callback, Closure $onResolved = null): Closure;
+    public function async(Closure $callback, Closure $onResolve = null): Closure;
 
     public function delay(float $seconds): void;