+++ docs/src/content/en/docs/getting-started/changelog.md
@@ -18,7 +18,10 @@ toc: true
 - Upgrade to Symfony 6.3
 - Refactor docs structure
 - Refactor tooling from https://github.com/jolicode/castor
-- Update `tick` and Remove `start` and `stop` from Flow\DriverInterface
+- Rework on Flow\DriverInterface
+  - Update `async` that now $onResolve get called with async $callback result or Flow\Exception as first argument
+  - Update `tick` that now return a Closure to cleanup tick interval
+  - Remove `start` and `stop`
 - Add Fiber Driver from https://github.com/jolicode/castor/blob/main/src/functions.php
 
 ### 1.1.2